refactor: 迁移实体获取模块至 `utils/`
明确其作为共享工具的职责

- `core/queries` → `utils/queries.ts`
- 更新12个文件的导入路径
