+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './parse-coordinates';
+export * from './queries';
